Move utils to root directory
